@@ -44,90 +44,91 @@
     "lint:md": "eslint -f codeframe  --ignore-pattern='*.js' --ignore-pattern='*.ts' --ignore-pattern='*.tsx' .",
     "lint:prettier": "pnpm run:prettier -- --check",
     "lint:repo": "manypkg check",
-    "refresh": "pnpm clean:all; pnpm symlink:root; pnpm install;",
+    "refresh": "pnpm clean:all; pnpm symlink:root; pnpm install -r;",
     "prerelease": "pnpm if-not-ci run-s checks:release",
     "release": "pnpm if-publishable run-s release",
     "release:latest": "pnpm publish -r",
-    "reset": "pnpm clean:all; rm pnpm-lock.yaml; pnpm symlink:root; pnpm install;",
+    "reset": "pnpm clean:all; rm pnpm-lock.yaml; pnpm symlink:root; pnpm i -r;",
     "run:prettier": "prettier --ignore-path .eslintignore \"**/*.{js,jsx,ts,tsx,md,mdx,json,html,css,yml,yaml,graphql}\"",
     "size": "size-limit",
     "storybook": "manypkg run support/storybook storybook",
-    "symlink:root": "node support/scripts/symlink-root.js",
+    "symlink:root": "node support/scripts/symlink-root.mjs",
     "test": "jest --verbose",
     "test:build": "cross-env TEST_BUILD=true jest --verbose --coverage=false",
     "test:watch": "jest  --watch --verbose=false --coverage=false",
-    "ts": "ts-node -P support/tsconfig.base.json -O '{\"module\":\"commonjs\"}' --transpile-only",
+    "ts": "node -r esm -r ./support/scripts/babel-register.js",
     "typecheck": "tsc -b ./tsconfig.references.json",
-    "update:repo": "pnpm update --latest --recursive -i",
+    "update:deps": "pnpm update --latest --recursive -i",
     "update:workspace": "pnpm up -r --workspace \"@remirror/*\"",
     "watch": "preconstruct watch # not recommended"
   },
   "browserslist": [
-    "since 2017"
+    "since 2018"
   ],
   "dependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/parser": "^7.11.5",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-proposal-decorators": "^7.10.5",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.10.4",
-    "@babel/plugin-proposal-numeric-separator": "^7.10.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.11.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.11.0",
-    "@babel/plugin-proposal-private-methods": "^7.10.4",
+    "@babel/core": "^7.12.3",
+    "@babel/parser": "^7.12.5",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/plugin-proposal-decorators": "^7.12.1",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.12.1",
+    "@babel/plugin-proposal-numeric-separator": "^7.12.5",
+    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
+    "@babel/plugin-proposal-optional-chaining": "^7.12.1",
+    "@babel/plugin-proposal-private-methods": "^7.12.1",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.11.5",
-    "@babel/plugin-transform-template-literals": "^7.10.5",
-    "@babel/plugin-transform-typescript": "^7.11.0",
-    "@babel/preset-env": "^7.11.5",
-    "@babel/runtime": "^7.11.0",
-    "@changesets/cli": "^2.10.2",
-    "@jest/types": "^26.3.0",
-    "@manypkg/cli": "^0.16.0",
+    "@babel/plugin-transform-runtime": "^7.12.1",
+    "@babel/plugin-transform-template-literals": "^7.12.1",
+    "@babel/plugin-transform-typescript": "^7.12.1",
+    "@babel/preset-env": "^7.12.1",
+    "@babel/runtime": "^7.12.5",
+    "@changesets/cli": "^2.11.2",
+    "@jest/types": "^26.6.2",
+    "@manypkg/cli": "^0.16.1",
     "@manypkg/get-packages": "^1.1.1",
-    "@preconstruct/cli": "^1.1.27",
-    "@size-limit/preset-big-lib": "^4.6.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@types/eslint": "^7.2.2",
-    "@types/jest": "^26.0.13",
-    "@types/jest-axe": "^3.5.0",
-    "@types/node": "^14.6.4",
-    "@types/testing-library__jest-dom": "^5.9.2",
-    "@typescript-eslint/eslint-plugin": "^4.1.0",
-    "@typescript-eslint/parser": "^4.1.0",
-    "babel-jest": "^26.3.0",
+    "@preconstruct/cli": "^1.2.0",
+    "@size-limit/preset-big-lib": "^4.7.0",
+    "@testing-library/jest-dom": "^5.11.6",
+    "@types/eslint": "^7.2.5",
+    "@types/jest": "^26.0.15",
+    "@types/jest-axe": "^3.5.1",
+    "@types/node": "^14.14.7",
+    "@types/testing-library__jest-dom": "^5.9.5",
+    "@typescript-eslint/eslint-plugin": "^4.8.1",
+    "@typescript-eslint/parser": "^4.8.1",
+    "babel-jest": "^26.6.3",
     "babel-plugin-annotate-pure-calls": "^0.4.0",
     "babel-plugin-dev-expression": "^0.2.2",
     "babel-plugin-macros": "^2.8.0",
     "cpy-cli": "^3.1.1",
     "cross-env": "^7.0.2",
-    "eslint": "^7.8.1",
-    "eslint-config-prettier": "^6.11.0",
+    "eslint": "^7.13.0",
+    "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-graphql": "^4.0.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jest": "^24.0.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jest": "^24.1.3",
     "eslint-plugin-jest-formatting": "^2.0.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-markdown": "^2.0.0-rc.0",
     "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-simple-import-sort": "^5.0.3",
+    "eslint-plugin-simple-import-sort": "^6.0.0",
     "eslint-plugin-sonarjs": "^0.5.0",
-    "eslint-plugin-unicorn": "^21.0.0",
+    "eslint-plugin-unicorn": "^23.0.0",
+    "esm": "^3.2.25",
     "husky": "^4.3.0",
-    "jest": "^26.4.2",
-    "jest-circus": "^26.4.2",
+    "jest": "^26.6.3",
+    "jest-circus": "^26.6.3",
     "jest-extended": "^0.11.5",
     "jest-watch-typeahead": "^0.6.1",
-    "lint-staged": "^10.3.0",
+    "lint-staged": "^10.5.1",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.1.1",
-    "prettier-plugin-packagejson": "^2.2.5",
+    "prettier": "^2.1.2",
+    "prettier-plugin-packagejson": "^2.2.7",
     "rimraf": "^3.0.2",
-    "size-limit": "^4.6.0",
+    "size-limit": "^4.7.0",
     "snapshot-diff": "^0.8.1",
     "ts-node": "^9.0.0",
-    "typescript": "^4.0.2",
+    "typescript": "^4.1.1-rc",
     "typescript-snapshots-plugin": "^1.7.0"
   },
   "engines": {
