@@ -1,15 +1,17 @@
 /**
+ * @script
+ *
  * This script edits the `/.changeset/config.json` file so that it doesn't need
  * the GITHUB_TOKEN to be available in order to build the changeset.
  *
  * This allows PR's which don't have access to the `GITHUB_TOKEN` to still pass.
- *
- * @packageDocumentation
  */
 
-const writeJson = require('write-json-file');
-const chalk = require('chalk');
-const { baseDir } = require('./helpers');
+import chalk from 'chalk';
+import writeJson from 'write-json-file';
+
+import changesetConfig from '../../.changeset/config.json';
+import { baseDir } from './helpers';
 
 async function main() {
   if (!process.env.CI) {
@@ -20,8 +22,7 @@ async function main() {
     return;
   }
 
-  const changesetConfig = require('../../.changeset/config.json');
-  delete changesetConfig.changelog;
+  Reflect.deleteProperty(changesetConfig, 'changelog');
   await writeJson(baseDir('.changeset', 'config.json'), changesetConfig);
 
   console.log(chalk`{green Successfully updated the CI configuration. }`);
