@@ -1,4 +1,10 @@
-const { execSync } = require('child_process');
+/**
+ * @script
+ *
+ * Run the command when not within a `CI` environment.
+ */
+
+import { execSync } from 'child_process';
 
 const [, , ...args] = process.argv;
 const command = args.join(' ');
